MB-44481 Set manifestUid to invalid value if collection is not available
In a mixed-cluster mode (e.g. 7.0 + 6.6 nodes) collection is disabled,
however in the query node (7.0) we leave the manefestUid to the default
value of 0. Since 0 is a valid manefestUid, when the legacy node (6.6)
is removed from the cluster the query node cannot properly detect that
now a valid manifestUid of 0 is being sent which represent a change. As
a result the query node remained in collections-disabled node until the
manefistUid changed for a separate reason.

To fix, use MaxUint64 to represent an invalid manifestUid, and when
collection is disabled, set manefistUid to MaxUint64. Later on when
collection becomes enabled, the query node can then detect a change
(from MaxUint64 to 0) and refresh its info properly.

Change-Id: I5440cff96eaa3204c57db560eb4964e560d009dc
Reviewed-on: http://review.couchbase.org/c/query/+/147029
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>