MB-40931 add node load to stats
The node load is not only useful for load balancing, but for sizing and auto provisioning as well.
So, report it in the stats as a gauge.
There's no point in adding a load gauge to the accounting metrics, because it would only add complexity, and it would only report the load at the time the last request terminated.
Having local server stats allows to report the load at the time the stats were requested.

Change-Id: I423ffac253e164f5dd33a14af95ed2f536bfee6b
Reviewed-on: http://review.couchbase.org/c/query/+/147487
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>