MB-44770 Proper handling of selectivities of range pairs
The optimizer currently assumes independence of multiple predicates,
and multiply selectivities of the predicates. When we have a range pair,
e.g., c1 >= 0 and c1 < 100, then we should NOT just multiply the
selectivities of the two predicates. Currently if c1 is used as index
spans we do properly handle the range pairs in terms of selectivity
calculation, however if c1 is not part of index spans then it's not
handled properly.

Move the logic for handling selectivities of range pairs to when
filers are classified and initial selectivities are calculated. Since
all predicates for each keyspace is in the filters array for that
keyspace, we just range through the filters and look for range pairs.
Currently we only do this for single-keyspace filters and not
join filters.

Also add logic to handle range pairs inside SATISFIES clause of ANY
predicates.

Change-Id: I43239dde2ed40ab86fe65caa5e6b49577f9bbe56
Reviewed-on: http://review.couchbase.org/c/query/+/147828
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>