MB-44759 Pick shorter index in case everything else is the same
In this defect we have two indexes, idx6 and idx7, which has exactly
the same set of index keys, except that idx7 has one extra index key
at the very end.

When we have a query that has predicates on key positions 0 and 2,
the index key information for idx6 and idx7 is exactly the same in
terms of sargKeys, minKeys, maxKeys, etc. Current code will then
pick the first index considered (which could be idx6 or idx7
depending on the order we range through a map).

Adjust the logic for comparing indexes such that in case of exactly
the same index key information, don't return, instead move on to
other comparisons such as index condition and index key lengths.
Also adjust the order of the last two checks such that in case of
equivalent indexes or almost equivalent indexes we return true
as the last check (when index length is also the same). This is to
avoid recurrence of MB-42039 and MB-42187.

Change-Id: I6cf70c7861b760a5754b85a817afbc9fb6b76b76
Reviewed-on: http://review.couchbase.org/c/query/+/147832
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>