MB-44944 intersect scan hangs on timeout
The issue is a race condition between the timeout stop and the first child terminating.
When the first child terminates, and then the timeout is received, we wait for all children,
but we had already successfully detected that the first child had gone, so we wait one time too many.
We had already started to deal with missed signals in 6.6.1, but it turns out there's more.
The issue can be traced back to at least Vulcan.

Change-Id: Ic5fe08ad05f047cc7daaaded9835482f47b75386
Reviewed-on: http://review.couchbase.org/c/query/+/148493
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>