MB-45272 Favor UNION SCAN over PRIMARY SCAN
When the query has OR condition, we can either use a single
index scan, or a UNION scan.

Modify the condition such that if the single index scan is
a PRIMARY scan, favor the UNION scan. Also, add a new function
in SecondaryScan interface, to return the index pointer in
case a single index is used underneath, or nil if multiple
indexes are used. In case of OR predicate, if a UNION scan
uses a single index underneath (i.e. same index used in
multiple arms of the OR), then don't use UNION scan.

Change-Id: I3fe2af2b89dec045bd37868a50eae04abcf5cf4f
Reviewed-on: http://review.couchbase.org/c/query/+/149672
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>