MB-46183 Set proper formalizer flag for index keys and conditions
When formalizing index keys and conditions a special flag is required.
This is done properly when indexes are being considered for index scans,
however similar logic is missing when indexes are considered for pattern
indexes. Add the same logic there as well. Also look for other places
where index keys and index condition is formalized and apply the same
change.

Change-Id: I3f5a1527ecb55faaf51db313c950d8f8c2399288
Reviewed-on: http://review.couchbase.org/c/query/+/153178
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-on: http://review.couchbase.org/c/query/+/153179
Well-Formed: Build Bot <build@couchbase.com>