MB-46284 Don't discard existing keyspace entry.
If all that is needed is a manifest refresh, don't discard the entry
and create a new one with new bucket-updater go routine.  When we do
this we leak the go routine and its connection to ns-server.

During an upgrade the GET_COLLECTIONS_MANIFEST message may fail (no
collections for the bucket) and we don't want to repeatedly recycle
the entry (leaking go routines) in this situation.

Change-Id: Ia17ae3dae35da08f5b6b8ee7009e37189ece6b39
Reviewed-on: http://review.couchbase.org/c/query/+/153789
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>