MB-46607 Avoid using WHERE clause filters to generate index spans for inner of LEFT OUTER JOIN
When a LEFT OUTER JOIN is being done, the WHERE clause is semantically evaluated
after the join, and for index selection purpose for the inner keyspace, we should
only use the ON-clause filter for index span generation. The previous fix for
MB-38564 partially addressed this issue but it's incomplete. When individual filters
are used for index span generation we should skip WHERE clause filters in case
of inner keyspace of LEFT OUTER JOIN.

Change-Id: I2953792240549dfc4a15291609682147563e4db1
Reviewed-on: http://review.couchbase.org/c/query/+/154597
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>