MB-38824 Add ARRAY_CONTAINS_ALL and ARRAY_CONTAINS_ANY functions.
ARRAY_CONTAINS, if passed an array as the second argument will look for the exact array in the first argument.
        ([1,2,3],[2]) -> false, ([1,[2],3],[2]) -> true

ARRAY_CONTAINS_ANY tests for any element in the second array being present in the first array.
        ([1,2,3],[2,9]) -> true

ARRAY_CONTAINS_ALL tests for all elements in the second array being present in the first array.
        ([1,2,3],[2,9]) -> false

Change-Id: If2bea55f6c16fcf117a7134b0f807980d599207d
Reviewed-on: http://review.couchbase.org/c/query/+/154827
Tested-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>