MB-47155 Keep track of UNNEST keyspaces differently
Currently unnest scan is only considered for the "primary" keyspace,
which is the first keyspace specified in join ordering. With join
enumeration, the concept of "primary" keyspace is out of date since
any keyspace (as long as it's not a subservient keyspace in an outer
join, or the right-hand side of a NEST operation) can be considered
a "primary" keyspace. Thus we need a different way to keep track of
UNNESTs.

Currently a keyspace is marked with a flag on its corresponding
baseKeyspace to indicate it's part of a (chain of) UNNEST operation.
This is going to be cumbersome if multiple keyspaces can now be
considered "primary" keyspace.

Instead, we already added a map of UNNEST aliases for the "primary"
keyspace as part of a previous bug fix. We can now just use that
mechanism instead of marking baseKeyspace as "primary unnest".

As part of the fix, the flag on baseKeyspace is changed from
marking "primary" unnest to marking any unnest, since it's still
beneficial to know whether the keyspace is part of an UNNEST.

Change-Id: I699aa1b5515be4779c180a34038f2b9985010889
Reviewed-on: http://review.couchbase.org/c/query/+/156603
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>