MB-46641 Don't use intersect scan for unnest scan when unnest scan subsumes the other index scan
When it is possible to do both a regular index scan and an unnest scan,
currently we always to an intersect scan using both indexes. This is
unnecessary when the unnest scan subsumes the regular index scan.

After an unnest scan is generated, check whether the unnest scan
subsumes an existing secondary index scan, if so, replace the secondary
index scan instead of doing an intersect scan with both.

Change-Id: I66e2d16aa7a02a9b73073054d648310fa2ca261d
Reviewed-on: http://review.couchbase.org/c/query/+/156718
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-on: http://review.couchbase.org/c/query/+/156719