MB-45296 Alias operator should inherit scope value from parent
Currently an Alias operator at runtime just keeps the values from
the subquery, but not inheriting any scope values from "parent".
This could be problematic if for example the query has a WITH
clause, and the values for the result of the WITH clause get
lost at the Alias step. This is only problematic when the alias
is for a subquery as the primary term of the query, since in a
join the "parent" values gets inherited from the left-hand side
of the join already.

The fix has a couple of aspects:

1. mark the primary term of a query
2. with a Alias operator, if it is for a primary term subquery,
   properly construct a inherited scope value

Change-Id: I3ffa20edd703845672f5ad5ee978a927a7317914
Reviewed-on: http://review.couchbase.org/c/query/+/156730
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>