MB-47356 Properly detect error for mixing ANSI join syntax with legacy join syntax
We do not allow mixing of ANSI join syntax with legacy join syntax
in the same query block. However this error condition is not always
properly detected, if there exists an UNNEST in between the ANSI join
and legacy join syntax.

Add logic to skip UNNEST when checking for mixing of join types.

For the new comma-separated join syntax, also move the error checking
for mixing of join type from parser to semantics, since it has the
same issue (and since we use ANSI join internally for comma-separated
join, it'll be caught by existing checks).

Change-Id: I2bbf0065c1bab3814b6765b447a903cf8b4e1d1e
Reviewed-on: http://review.couchbase.org/c/query/+/157374
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>