MB-47357 tiny race condition between child terminating and parent waiting for more data
This needs to be tested in longevity properly, but the issue with the intersect scan waiting forever
is yet a corner case in the dance between parents killing and children terminating.
We have to have yet another state where a child knows it's still alive, and therefore has to alert the parent of its termination,
and once we switch from that to killed, we have to alert the parent in its stead.

Change-Id: Ia5c1176e3665a72606b89c9cc79bb8ccbb456b8c
Reviewed-on: http://review.couchbase.org/c/query/+/157440
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>