MB-45984 Allow ON clause to be nil for comma-separated join
Comma-separated join is implemented using ANSI JOIN, and
currently a TRUE expression is used as ON clause to fit
existing code (since currently code assumes a non-nil
ON clause for ANSI JOIN).

It might be useful later on to know whether an ANSI JOIN
is originally a comma-separated join, thus we add a new
flag on simple-from-term to mark whether a comma-separated
join syntax is being used. Also, instead of using a TRUE
expression (which seems redundant), set ON clause to be nil,
and fix the assumptions in the code such that nil ON clause
can be properly handled.

Change-Id: I3b39946f3d07a23c4a528d5901bc6dfce04aa81c
Reviewed-on: http://review.couchbase.org/c/query/+/157459
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>