MB-47547 Properly fix up build/probe side of hash join if CBO switches them
When building hash join, the build/probe sides of the hash join need to be
properly setup. The issue in this bug is that CBO may switch the build and
probe sides of the hash join based on cost information, and part of the
setup is already done, leading to mismatch of build/probe expressions and
the rest of the runtime operators.

The fix is to delay the setup of build/probe sides of hash join until
after potential CBO switch. (This is already fixed in 7.0.0 and later).

Change-Id: Ie46e5aa03e70fc76db722477cee8cf8d6e66d481
Reviewed-on: http://review.couchbase.org/c/query/+/158391
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>