MB-26606 Add parentheses to IN and WITHIN syntax.
This harmonises IN and WITHIN syntax with common SQL/RDBMS syntax by
allowing the IN-list to appear in parentheses.

The primary challenge with this implementation is to differentiate
between (expr) and (single-value-in-list), since in-list elements are expressions
too.  This forces some run-time processing in IN and WITHIN evaluation.

This handles most cases making () equivalent to [], but there are
bound to be some edge cases where it isn't identical.

(Note: we support an empty array here via both () and [].)

Some queries will have to change - and I think this is the death knell
for this MB; e.g.

select 1 from {"a":1,"b":1} t
where t.a in (case when t.c is valued then [t.b] else [1,2] end);

now needs to be written as:

select 1 from {"a":1,"b":1} t
where t.a in case when t.c is valued then [t.b] else [1,2] end;

i.e. the CASE needs to not be parenthesised in order to work - so the IN list of
e.g. [1,2] is not seen as [[1,2]].

We could be more selective when applying () as array delimiters, but
there will always be a trade-off and it is surely clearer to always
apply () after IN/WITHIN as meaning [].

Change-Id: I876bcda63e1e5f455ff9ae37abe30b9479957fbb
Reviewed-on: http://review.couchbase.org/c/query/+/158641
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>