MB-47820 Use specific error code for "no index available"
Currently when there is no index available for a keyspace, an error
is returned via fmt.Errorf(). This results in a generic error code
of 4000 (plan error) being returned. For join enumeration, we need
to distinguish this specific case where no index is available from
other error cases, since "no index available" is not necessarily a
fatal error, since we could still potentially generate an index
path later on when join predicates are considered.

We need to mark this error case with a specific error code such
that it can be checked. A new error code is used for this error.
However, to maintain backward compatibility, we wrap the new error
inside a generic plan error (such that the same error code 4000
and the same error message is returned), however we can look
(if necessary) inside the "cause" of the generic error to find
the new specific error code.

Change-Id: Ie73d5b1d78a9a5167d23aa0b73b567efb97395d6
Reviewed-on: http://review.couchbase.org/c/query/+/158970
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>