MB-47938 Better handling of nested array key in CBO
The current logic in CBO handles one-level array index key
properly, but if there is a nested array index key there
are several places where the nesting is not handled properly.

For each baseKeyspace, a map exists such that for an UNNEST
scan, we can find which UNNEST it is for. We'll need to
enhance the UNNEST alias to be an array of aliases for the
multiple level of unnests.

Also when generating a standardized ALL expression to use
in histograms, handle nested array keys as well.

Change-Id: I2991930f8e89d847fd55bbe79f3726e78bd79861
Reviewed-on: http://review.couchbase.org/c/query/+/159332
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-on: http://review.couchbase.org/c/query/+/159337