MB-48266 Replace query parameters for WHEN clause of ANY predicates
This defect is somewhat similar to MB-42956, and is a regression
from MB-33009.

The fix for MB-33009 delayed replacement of query parameters during
query planning. For this defect, a query parameter is used inside
the SATISFIED clause of an ANY predicate, and this should have
matched the WHEN clause of an array index key, however due to the
presence of query parameters the match is not properly detected.

To fix this, we need to do query parameter replacement first before
matching to the WHEN clause. This requires the PrepareContext which
is added to a couple of necessary structures.

Change-Id: I6c53f77f4139f5715b13e10d4e54d158f7b2ca05
Reviewed-on: http://review.couchbase.org/c/query/+/160668
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-on: http://review.couchbase.org/c/query/+/160669
Well-Formed: Restriction Checker