MB-19784 Permit dropping of named primary index via DROP PRIMARY INDEX statement
This allows the sequence:

CREATE PRIMARY INDEX myprimary ON default;
DROP PRIMARY INDEX myprimary ON default;

but won't permit:

CREATE INDEX myindex ON default(a);
DROP PRIMARY INDEX myindex ON default;

(must be a primary index in the DROP PRIMARY statement)

Existing syntax is unaffected and a primary index can still be dropped
with a plain DROP INDEX statement.  The indention is only to simplify
use (i.e. only single word change needed).

(It is in effect an extension of MB-31295.)

Change-Id: I14c6cc567288805ad05c4a1c8309825c475bc08e
Reviewed-on: http://review.couchbase.org/c/query/+/160710
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>