MB-40533 Revisions to handle syncronisation issues
These changes are to handle the situation where we are told of topology
changes before the cluster information has been updated.  They have us
delay the PrepareTopologyChange in order to retry.  We'll also try fill
in any gaps whenever we're asked for our topology and again before
attempting to eject any nodes.

Change-Id: Id44646cb5e5a213c56cc4f86d502341d245b5580
Reviewed-on: http://review.couchbase.org/c/query/+/160774
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>