MB-47175: Support CAFile option from ns server
NS Server wants to change (or fix) how they use certs.
certfile, keyfile - these used to rep the certificate and key but with 7.1 it'll rep the chain and key
with the new option CAfile rep'ing the root ca.pem.

For connections to KV the client (us) will load the cert using the new option ca.pem (caFile)
For listeners, we will use the chain and key (cerfile and keyfile) to load the cert and then append the new cert to that.

Change-Id: I4e033d27276380058e68eb97817e41056183e5ca
Reviewed-on: http://review.couchbase.org/c/query/+/161334
Tested-by: Isha Kandaswamy <isha@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>