MB-48545 Support specifying optimizer hints as comment in a query block
Add support for specifying optimizer hints as a comment in a query block.
For now this is only supported for SELECT statement, and the comment
be right after the SELECT keyword. A plus sign must immediately follow
the opening of a block comment or line comment, like:

  /*+ INDEX(keyspace1 ix1 ix2) USE_HASH(keyspace2 keyspace3/BUILD) ORDERED */
or
  --+ INDEX(keyspace1 ix1 ix2) USE_NL(keyspace2 keyspace3)

You can also specify the hints as a JSON object, like:

  /*+ {"index": {"keyspace": "keyspace1", "indexes": ["ix1", "ix2"]}, "use_hash": {"keyspace": "keyspace2", "option": "BUILD"}} */
or
  --+ {"index": {"keyspace": "keyspace1", "indexes": "ix1"}, "use_nl": [ {"keyspace": "keyspace2"}, {"keyspace": "keyspace3} ] }

The hints that's specific to a keyspace: INDEX, USE_NL, USE_HASH, when
specified in the new style, are "converted" to existing style such that
they can be used properly in existing code. The other kinds of hints
(ORDERED) will be handled separately and not in this checkin.

We do NOT allow mixing of old style hint and new style hint specified for
the same keyspace. Also the hint block must be in either all JSON format
or all relational format.

The explain will be updated in a later checkin to reflect the hints
specified, for now there is no indication of the hints specified or
whether the hints are followed or not (which is the case currently
for the old style hints).

Change-Id: I614c44d842d31c665121c1cc403dcc590056f2d4
Reviewed-on: http://review.couchbase.org/c/query/+/161989
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>