MB-48027 Add LEN() function as a general-purpose length function
Before this, you'd have to use LENGTH with strings, ARRAY_LENGTH
with arrays and OBJECT_LENGTH with objects.  This just rolls them into
one that returns the length based on input type.  LEN so as not to
conflict with any possible reliance on error return from existing
string-only LENGTH function.

Change-Id: I670a98586acb5e77af53bf35e90af11922daecb4
Reviewed-on: http://review.couchbase.org/c/query/+/162461
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>