MB-48617 Add fallback when random document retrieval fails
When creating a random document retriever we'll try a number (5) of times to
obtain a document key.  If we're unable to, we'll fail the retriever
creation falling back to an index retriever.

In addition to and tried after the primary index retriever, an "any
GSI index" retriever has been added which will order indexes preferring
unconditional, non-array with the shortest key.  These will be used in
listed order to retrieve the documents until the sample count is reached (or
there are no further unique keys).  With either index retriever there is
no randomness to the document selection itself.

When the primary index is used, a warning, e.g.

    "warnings": [
        {
            "cause": {
                "indexes": [
                    "#primary"
                ]
            },
            "code": 7018,
            "msg": "Index scanning used; document sample may not be representative."
        }
    ],

is produced.  When other indexes are used, a warning, e.g.

    "warnings": [
        {
            "cause": {
                "indexes": [
                    "ix_one_field",
                    "ix_one_field_2",
                    "ix_two_field",
                    "ix_three_field",
                    "ix_array",
                    "ix_with_condition"
                ]
            },
            "code": 7018,
            "msg": "Index scanning used; document sample may not be representative."
        }
    ],

is produced; this lists the indexes used to fulfil the sample size
criterion.  If only a single index is used, only one will be listed.
This example shows the order of preference (embedded in the index
names).

Change-Id: Ib23f017c0774f659af4699bdb128e6f79fe08188
Reviewed-on: http://review.couchbase.org/c/query/+/162486
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>