MB-48734 Update document retrieval for inferencer and update stats.
This implements a "unified" document retriever that will attempt random
document selection, primary key (random) scanning or ranked secondary
index (random) scanning to provide documents for sampling.  It will
fall back to index (primary or secondary) if too many duplicates
are seen from the random document retrieval API.

An upper limit of 20000 (serial) scans is imposed for index key
retrieval.  The document count is divided by the number of scans to be
used providing logical blocks.  Scans randomly pick offsets in
sequential blocks to return the keys (for primary scsns, offsets are
relative to block size and based on the last key retrieved such that
spans are used to keep offsets supplied to the indexer relatively small
regardless of the data set size).  If the sample size is greater than
20000 then each scan will return multiple sequential keys starting at
the random offset noted.

Secondary scans implement randomness by reading and discarding a random
number of keys in each block.  The block size is calculated using the
index count and remaining samples each time an index scan is initially
used.  If a timeout occurs, sampling is resumed from the next index
key which can result in skipping large portions of highly duplicate
indexes.  Block alignment is reset to the start of the key but the size
isn't recalculated in order to prevent compressing sampling toward later
keys.  If the end of an index is reached before the sample size has been
met it is expected that subsequent indexes will fulfill the remainder.

Change-Id: I4c118552e45f535eea7f5358a77bd187569ef1a4
Reviewed-on: https://review.couchbase.org/c/query/+/162681
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>