MB-20895 Prune sort with partial index order.
When the input is partially sorted we can:
- discard distinct blocks of partially sorted items until we reach the
  specified offset
- fully sort and stream blocks of partially sorted items up to the
  specified limit
- halt processing once the limit has been reached

Over all this hopes to limit the memory consumption of sorting when
already partially sorted by an index.  (When fully sorted by an index
there is no order-by operator.)

In the case where all items have the same partial sorting index terms
and neither limit nor offset are specified, it should be no worse than
an order-by without any index pre-sorting.

Change-Id: I2d9456bb375dbd51c23ae6fdddce56bf2e35efa3
Reviewed-on: https://review.couchbase.org/c/query/+/164582
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>