MB-48658 Flush batch sooner.
execution.PipelineBatchSize() is decoupled from the batch pool element
size as a minimum size is now enforced to support possible increases in
the batch size and still make use of pooling.

The default batch pool element size is this minimum.

The operators will bypass batching if the downstream side is handling
the operator's output.  This is determined by looking at the output
queue length and if more than 50% empty items are passed on immediately;
if not then they're batched before being passed on.

The batch size will increase from the default up to the pipeline cap
(value exchange cap) if the input queue exceeds the current batch size.
This is incremental based on the queue length allowing for mas many
optimal variations as possible.  The size however is only ever increased
during execution but coupled with immediate flushing it doesn't
adversely impact after the peak.

Immediate flushing is to better support flows that benefit from a steadier
rate and dynamic batch size increases help where input to the operator
may be faster than it can consume, and downstream better deals with
larger batches.  The aim is the best of both worlds, of course.

Further, the batch element array is acquired once from the pool and
simply reset between batches as opposed to being released and acquired
for each batch.

Change-Id: I4237138e53dd35f515dd437e9ea574a844c6cdc3
Reviewed-on: http://review.couchbase.org/c/query/+/162900
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>