MB-46802 Abort non-transaction requests when the number of errors raised reaches a limit.
The limit is the "request-error-limit" setting that defaults to 16.
If this is set to 0, this processing is disabled.

When aborted, an additional error is added with details of unique &
duplicate error counts. This includes the mutation count to highlight
any DML successes, regardless of the inclusion of metrics.

Warning de-duplication has been moved from the output writing code to
the warning accumulation code in the request. (Potential memory savings
with immediate duplicate exclusion.)

Change-Id: Ic1a7bb067e5bc6495ab5b06a665d12ced7dabfc3
Reviewed-on: http://review.couchbase.org/c/query/+/163415
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>