MB-48732 amend retry count mechanism, set read from replica at the node level
The primary scan test fails because with the primitives retry standardization,
GetBulk() had changed from retrying 25 times, to retrying 2 x nodes, and the test in question was running on a single node.
Standardized retrying to 2xnodes across all fetches, but introduced a minimum and maximum number of retries.
Currently set at 25 & 100, but can be changed to whatever the consensus is.

Also, switch read from replica setting from bucket level to node level.

Change-Id: I8a139544dcd5bfd57a1f81f9433682633ba78687
Reviewed-on: http://review.couchbase.org/c/query/+/163660
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>