MB-49005 retry multiple times KV response on NMVB
The underlying issue here is that the vbmap sent by the KV might actually be the old one,
or even worse, it may be changing under KV's feet.
Retry the vbmap sent by KV on all attempts, not just the first one, and don't skip a KV get if the vbmap version hasn't changed:
the next get might give us the right map.

Change-Id: I1ad30e81ca769aa7e52000631ebe4ddacd8118ac
Reviewed-on: http://review.couchbase.org/c/query/+/164054
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>