Fix local build
Attempt to fix the build issues with recent jsevaluator change.

Sitaram suggested that we use the istalled server as library path
instead of $GOPATH/lib. This requires that the installed server
is "recent" but should give us less dependency on what the local
build has.

I had trouble trying to get the space in the path to work with
CGO_LDFLAGS on Mac, Sitaram suggested we use LIBRARY_PATH on Mac
which seems to work. I was able to get build and unit test working
on Mac, and build working on CentOS, but I've never tried running
unit tests on CentOS.

Marco/Donald if you can give this a try in your env and see
whether running unit tests works.

Please feel free to change or suggest a better solution.

Change-Id: I5fc315dae68fcc7376bca9a9cd1ef463400fd3bf
Reviewed-on: http://review.couchbase.org/c/query/+/164376
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>