MB-49362 marking and display of optimizer hints
When optimizer hints are specified, as we build the plan, either
with RBO or with CBO (with or without join enumeration), we need
to properly mark whether there are any errors with optimizer
hints (in terms of a hint not being followed). Once a plan is
chosen, we put information on optimizer hints in the explain
in several groups: hints that are followed, hints that are not
followed, hints with error, invalid hints, etc.

Change-Id: Ifb6fed46ad034167a596d9eed0c7b9fce7901494
Reviewed-on: http://review.couchbase.org/c/query/+/165048
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>