MB-49432: Ensure in-build library paths are first on link path
If we link against the install directory (which is where third-party
libraries are located) first, then when doing incremental builds during
development, we'll try to link against libraries from earlier builds
such as libjseval. If those libraries happen to have a breaking API
change, or introduce new APIs that goxdcr depends on, the build will
break with a strange link error.

Change-Id: I1af2c42cc71a4218fcb938489e598ca5b67baadd
Reviewed-on: http://review.couchbase.org/c/query/+/165314
Tested-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>