MB-49532 UDF execute prepared returns no results
This took me a while to debug, but it turns out that instantiated prepared statements by default have a stream operator, which eats the results before they get to the collector.
I have tried having the internal request Output to send the results to the collector, but things get complicated by the fact that Stream recycles results, which is not good.

Swapping Stream for one of the collectors seems a strategy that works.
Note that now for internal statements we have to position a stream too in order to make collector handling uniform across prepared and non prepared statements.

Change-Id: Id08fd9cf7cf1442301297bb3f4d9a7e4ed962a01
Reviewed-on: https://review.couchbase.org/c/query/+/166510
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>