MB-49823 cannot load travel sample
In the testcase, the loader uses libcouchbase with prepared statements and auto execute to create the indexes.
Internally, this uses ExecutePrepared() against the plan for the PREPARE CREATE INDEX...
Issue is, this plan doesn't have a Stream operator, so the Collect operator never gets to be executed.
Altered eval_stmt.go to swap the Stream operator if it's there, and add the Collect if it's not.

Change-Id: Ib10b72259168dee3c396865bd172bc2663db7ac1
Reviewed-on: https://review.couchbase.org/c/query/+/166779
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>