MB-51581 Proper marking of identifiers in filter covers when covered unnest scan is used
In the problematic query a join is being performed and the left-hand
side of the join uses a covered unnest scan. The array index used
has a WHEN clause. When we construct filter covers we include the
WHEN clause in the filter cover. However the expression is a newly
constructed expression and thus not having any of the previously
set flags (on the same expression used earlier). This may cause
issues later on when we consider index covering for the right-hand
side of the join, since the identifier for the binding variable
(which is the unnest alias) does not have proper flags set.

We already have logic to find and mark any keyspace identifiers
in the new expression. Enhance that logic to also find and mark
binding variables in the new expression.

Change-Id: I4172aa709c14c4f7a183169025d6df37d768268a
Reviewed-on: https://review.couchbase.org/c/query/+/166931
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/172923
Well-Formed: Restriction Checker