MB-49853 part 1 - remove heap allocation from metrics
In a specific CBSE worked by Donald he spotted a major bottleneck in the stats allocating buffers in order to work out the pecentiles.
Plus the request timer has always eaten about 3% in throughput accounting for each request.
The first part of the fix removes the heap allocation.

For the percentiles, couldn't really use a pool because
1) different timers might have different sample sizes
2) it would mean adding a destructor, which currently the metrics package doesn't offer
3) I didn't really want to pool every possible buffer allocated, but just keep a couple around in order not to bloat memory.

Change-Id: I57fffb40182c84d9a8b384c66b78fa045b04ef6d
Reviewed-on: https://review.couchbase.org/c/query/+/167048
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>