MB-49924 Missing SORT operator for query with ORDER BY with join enumeration
When a query has ORDER BY, the builder has initially set this.order to be
the order terms. Without join enumeration, this.order is reset to nil
during query planning process depending on many factors including joins,
index choices, etc.

With join enumeration, we use a copy of the builder during join enumeration.
As a result this.order on the original builder is never reset to nil. This
causes issues later on when we decide whether to build a SORT operator.

The fix is to reset this.order if the final plan chosen by join enumeration
does not provide ordering. Also reset this.limit and this.offset since we
only go through join enumeration when joins are present in the query in
which case index pushdowns are not performed.

Change-Id: Iae84a120d8d09334f7b9f32d2918b0db0018940b
Reviewed-on: https://review.couchbase.org/c/query/+/167071
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>